models/team: Move `Team::create_or_update_github_team()` fn into controller
This fn is doing way to much non-database-related stuff, so it shouldn't live in the `models` module.
